auth flow fixed
