+++ lib/views/feeds/feeds_view.dart
@@ -1,12 +1,76 @@
+import 'package:ecngnews/components/news_card_item.dart';
+import 'package:ecngnews/utils/ecng_theme.dart';
+import 'package:ecngnews/utils/size_config.dart';
+import 'package:ecngnews/views/feeds/feeds_viewmodel.dart';
+import 'package:ecngnews/views/home/home_view.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_shimmer/flutter_shimmer.dart';
+import 'package:stacked/stacked.dart';
 
 class FeedsView extends StatelessWidget {
-  final Widget child;
-  const FeedsView({Key key, this.child}) : super(key: key);
+  const FeedsView({
+    Key key,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return child;
+    return ViewModelBuilder<FeedViewModel>.reactive(
+        builder: (context, model, child) => Column(
+              children: [
+                //filters
+                model.newsCategory == null
+                    ? ListTileShimmer()
+                    : Container(
+                        padding: EdgeInsets.symmetric(
+                            vertical: SizeConfig.sizeMultiplier),
+                        child: ListView.builder(
+                          scrollDirection: Axis.horizontal,
+                          itemCount: model.newsCategory.length,
+                          itemBuilder: (BuildContext context, int index) {
+                            return CategoryCard(
+                              newsCategory: model.newsCategory[index],
+                            );
+                          },
+                        ),
+                        margin:
+                            EdgeInsets.only(bottom: SizeConfig.sizeMultiplier),
+                        height: SizeConfig.heightMultiplier * 16,
+                        decoration: BoxDecoration(
+                            borderRadius: BorderRadius.vertical(
+                                bottom: Radius.circular(
+                                    SizeConfig.sizeMultiplier * 2)),
+                            boxShadow: EcngColors.cardBoxShadow,
+                            color: Colors.white),
+                      ),
+                Expanded(
+                    child: Container(
+                  child: model.isBusy == null
+                      ? VideoShimmer()
+                      : model.appnews == null
+                          ? Container(
+                              child: Center(
+                                child: IconButton(
+                                  icon: Icon(Icons.refresh),
+                                  onPressed: model.listenToGeneralNews,
+                                ),
+                              ),
+                            )
+                          : ListView.builder(
+                              itemCount: model.appnews.length,
+                              itemBuilder: (context, index) =>
+                                  MoreNewsComponent(
+                                itemCreated: () {
+                                  if (index % 20 == 0) model.getMoreNews();
+                                },
+                                child: NewsItemCard(
+                                  news: model.appnews[index],
+                                ),
+                              ),
+                            ),
+                )),
+              ],
+            ),
+        viewModelBuilder: () => FeedViewModel());
   }
 }
