+++ lib/views/feeds/feeds_viewmodel.dart
@@ -1,3 +1,4 @@
+import 'package:ecngnews/models/news_category.dart';
 import 'package:ecngnews/models/news_model.dart';
 import 'package:ecngnews/services/news_service.dart';
 import 'package:ecngnews/utils/locator.dart';
@@ -5,32 +6,34 @@ import 'package:stacked/stacked.dart';
 
 class FeedViewModel extends StreamViewModel {
   NewsService _newsService = locator<NewsService>();
-  List<News> _news;
+  List<News> _news = List<News>();
   List<News> get appnews => _news;
-  List<String> _categories = [
-    'business',
-    'entertainment',
-    'general',
-    'health',
-    'science',
-    'sports',
-    'technology'
-  ];
-  List<String> get categoriesList => _categories;
-  Stream listenToNews() async* {
+  List<NewsCategory> _newsCategories = List<NewsCategory>();
+  List<NewsCategory> get newsCategory => _newsCategories;
+
+  Stream listenToGeneralNews() async* {
     _newsService.listenToGeneralNews().listen((event) {
       List<News> updatedNews = event;
-      if (updatedNews != null && updatedNews.length > 0) {
+      if (updatedNews != null) {
         _news = updatedNews;
-        notifyListeners();
+        // notifyListeners();
       }
     });
   }
 
-  void getMoreNews() => _newsService.requestMoreNews('general');
-  // @override
-  // Future futureToRun() => listToNews();
+  Future setNewCategoryPan() async {
+    _newsCategories = await _newsService.getCategory();
+    print('printing ');
+    notifyListeners();
+    return _newsCategories;
+  }
+
+  void getMoreNews() {
+    _newsService.requestMoreNews('general');
+  }
 
   @override
-  Stream get stream => listenToNews();
+  Stream get stream {
+    return listenToGeneralNews();
+  }
 }
